refactored db helpers
added logger
